Moved oidcClient requests to OAuth2.tx
